pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brandmon"
version = "0.1.0"
description = "BrandMon — Domain impersonation monitor (FastAPI + Jinja2/HTMX, PostgreSQL, Redis, Celery)"
authors = [{name="BrandMon Maintainers"}]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.35",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.13.2",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "httpx>=0.27.2",
  "dnspython>=2.6.1",
  "celery>=5.4.0",
  "redis>=5.0.8",
  "python-json-logger>=2.0.7",
  "structlog>=24.4.0",
  "websockets>=13.1",
  "email-validator>=2.2.0",
  "python-dateutil>=2.9.0.post0",
  "itsdangerous>=2.2.0"
]

[tool.setuptools.packages.find]
where = ["."]
